+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 
 Unreleased
 ==========
+* feat: Expand/collapse enabled for navigation tree in admin changelist view.
 * feat: Added delete confirmation template to overwrite delete view breadcrumbs
 * fix: Extended app Node deletion 404 bugfix